Yukti-API/requirements.txt

Django>=4.2,<5.0
djangorestframework>=3.14
python-decouple>=3.8
psycopg2-binary>=2.9
django-cors-headers>=4.3

dotenv